Extract email template settings into own module
Move the email-template save/preview handlers, their validation,
type guards, and preview sample data out of the oversized
settings.ts route file into a dedicated settings-email-templates.ts.
